web/requirements.txt
django==3.1.7
mysqlclient==2.0.3  # is mysql
django-adminlte3==0.1.6
django-environ==0.4.5
django-cors-headers==3.7.0
djangorestframework==3.12.2
pandas==1.2.4
django_cleanup==5.2.0
django-q==1.3.9
cx-Oracle==8.1.0    # if oracle
psycopg2==2.9.1     # is postgresql
js_regex==1.0.1
django-pwa==1.0.10
gunicorn==20.1.0